Rename cmds pkg to commands
This mnatches kinflate pkg
https://github.com/kubernetes/kubectl/tree/master/pkg/kinflate/commands
