+++ commands/up.go
@@ -1,4 +1,4 @@
-package cmds
+package commands
 
 import (
 	"io/ioutil"
