+++ commands/validate.go
@@ -1,4 +1,4 @@
-package cmds
+package commands
 
 import (
 	"fmt"
